[hardening] check object digests in get_balance_changes
This makes the results of the `balanceChanges` part of a tx effects more trustworthy by crashing unless the digest of coin objects in the effects matches the digest of coin objects we fetched from the local store. This way, the RPC provider and validators must be on the same page re: the value of coins.